Added helptext for System/User/UserUpdateWithSchoolyearChange and reworked smarty-gettext
Feature: smarty-gettext
renamed folder smarty3-i18n to smarty-gettext.
Smarty-gettext is now a Submodule of the git Repository and is a fork of the original smarty-i18n on github.
The parameter condense=yes was added, allowing formatting of text in the gettext-tags {t}{/t}.
All redundant whitespaces are removed automatically, so they do not appear later on when translating the strings.

Feature: Helptext for Module System/User/UserUpdateWithSchoolyearChange
A help displaying what columns the csv-file needs and how it should be structured.
Also added a Css-class .example-container, allowing to nicely highlight text like examples.