Adding new Components for improving Groups-usage
Added classes GroupsManager and UserGroupsManager
Allows for easier retrieving of groups.

Created new namespace-convention
Components not being in a particular Subprogram should have a namespace-path beginning with
Babesk
and continuing with normal module-names.
For example, groupsManager and UserGroupsManager have the namespace
Babesk\System
because Groups-Management belongs to the System-Module (They can of course be used in other modules, too)

Added function nestedSetToArray() to class ArrayFunctions.

Added some interfacing with the childs of class Group.