Refactoring and bugfix to ShowBooklist.php
ShowBooklist is still buggy with calculating the amount of books needed.
The whole system needs to be just refactored, it slows the calculation down and
its not easy to work out how it works.
Made ShowBooklists fetching of inventory-data of books a bit faster.
It also now displays the amount of books generally needed for lending.