Bulk of changes of RechargeCard, mostly finished
Modal works now. It displays a nice inputfield.
Selectors! Use Shift-1 to Shift-0 to select users without a mouse.
Intelligent focusing, so you just have to roll your face on the keyboard to do
all the fun registration stuff.
Now shows credits and updates them automatically.
The first JS-Code to use HTTP-Response-Codes! I am planning to expand on it,
since it is the normally used method in the WWW.